backend/requirements.txt
Django>=4.2.0,<5.0.0
django-cors-headers>=4.3.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
python-dotenv>=1.0.0
scrapy>=2.11.0
mysqlclient>=2.2.0; sys_platform != 'win32'
# For Windows, you might need to install mysqlclient manually 